Resolve Review Comments
