+++ src/gpuhunt/providers/vultr.py
@@ -5,8 +5,13 @@ import requests
 from requests import Response
 
 from gpuhunt import QueryFilter, RawCatalogItem
-from gpuhunt._internal.constraints import KNOWN_AMD_GPUS, KNOWN_NVIDIA_GPUS, is_nvidia_superchip
-from gpuhunt._internal.models import AcceleratorVendor, CPUArchitecture
+from gpuhunt._internal.constraints import (
+    KNOWN_AMD_GPUS,
+    KNOWN_NVIDIA_GPUS,
+    get_gpu_vendor,
+    is_nvidia_superchip,
+)
+from gpuhunt._internal.models import CPUArchitecture
 from gpuhunt.providers import AbstractProvider
 
 logger = logging.getLogger(__name__)
@@ -165,18 +170,6 @@ def get_gpu_memory(gpu_name: str) -> Optional[int]:
         if gpu.name.upper() == gpu_name.upper():
             return gpu.memory
     logger.warning(f"Unknown GPU {gpu_name}")
-    return None
-
-
-def get_gpu_vendor(gpu_name: Optional[str]) -> Optional[str]:
-    if gpu_name is None:
-        return None
-    for gpu in KNOWN_NVIDIA_GPUS:
-        if gpu.name.upper() == gpu_name.upper():
-            return AcceleratorVendor.NVIDIA.value
-    for gpu in KNOWN_AMD_GPUS:
-        if gpu.name.upper() == gpu_name.upper():
-            return AcceleratorVendor.AMD.value
     return None
